Add hive support(3)
